requirements.txt
discord.py==1.7.3
python-dotenv
aiomysql
numpy
matplotlib
pymongo
scipy
mysql-connector-python
requests
websockets